[refactor] add useKeywordInfo hook
